consolidate helper functions
